Fixed a test failing because i was trying to delete a venue with no seat configurations. Justed garded against the condition

+++ src/banian/models.py
@@ -125,7 +125,8 @@ class SeatConfiguration(db.Model):
         if not self.mutable():
             raise AssertionError
         seat_groups = self.seatgroup_set
-        db.delete(seat_groups)
+        if (seat_groups): 
+            db.delete(seat_groups)
         super(SeatConfiguration,self).delete()
 
     def put(self):
